requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=8.0.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.60.0
